modernize lazy-static infra
