+++ .gitignore
@@ -2,3 +2,4 @@ target
 doc
 Cargo.lock
 .cargo
+wip
