--- .travis.yml
@@ -1,36 +1,0 @@
-language: rust
-matrix:
-  include:
-    - rust: 1.40.0
-      script:
-        - cargo test --tests
-    - rust: stable
-      script:
-        - cargo test
-        - cargo test --features spin_no_std
-    - os: osx
-    - rust: beta
-    - rust: nightly
-      script:
-        - cargo test
-        - cargo bench
-        - cargo test --features spin_no_std
-        - cargo bench --features spin_no_std
-        - cd compiletest
-        - cargo clean
-        - cargo test
-        - cd ../
-
-    - rust: nightly
-      before_script:
-        - pip install 'travis-cargo<0.2' --user && export PATH=$HOME/.local/bin:$PATH
-      script:
-        - cargo doc --no-deps --all-features
-      after_success:
-        - travis-cargo --only nightly doc-upload
-script:
-  - cargo test
-
-env:
-  global:
-  - secure: YXu24LptjeYirjWYjWGsMT2m3mB7LvQATE6TVo7VEUXv8GYoy2ORIHD83PeImxC93MmZ01QeUezRzuCW51ZcK92VnNSBttlF60SvIX18VsJrV92tsAhievFstqYQ+fB8DIuQ8noU0jPz7GpI+R9dlTRSImAqWOnVIghA+Wzz7Js=
