--- compiletest/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "lazy_static_compiletest"
-version = "0.0.1"
-publish = false
-authors = ["lazy_static contributors"]
-
-[dependencies.lazy_static]
-path = "../"
-
-[dependencies.compiletest_rs]
-version = "0.3"
