+++ tests/compile_fail/incorrect_visibility_restriction.rs
@@ -1,0 +1,8 @@
+#[macro_use]
+extern crate lazy_static;
+
+lazy_static! {
+    pub(nonsense) static ref WRONG: () = ();
+}
+
+fn main() { }
