+++ tests/compile_fail/static_is_private.stderr
@@ -1,0 +1,14 @@
+error[E0603]: static `FOO` is private
+  --> tests/compile_fail/static_is_private.rs:13:31
+   |
+13 |     assert_eq!(*outer::inner::FOO, ());
+   |                               ^^^ private static
+   |
+note: the static `FOO` is defined here
+  --> tests/compile_fail/static_is_private.rs:6:9
+   |
+6  | /         lazy_static! {
+7  | |             pub(in outer) static ref FOO: () = ();
+8  | |         }
+   | |_________^
+   = note: this error originates in the macro `lazy_static` (in Nightly builds, run with -Z macro-backtrace for more info)
