+++ tests/compile_fail/static_is_sized.rs
@@ -1,0 +1,9 @@
+#[macro_use]
+extern crate lazy_static;
+
+lazy_static! {
+    pub static ref FOO: str = panic!();
+}
+
+
+fn main() { }
